# openapi/synapse.yaml
openapi: 3.1.0
info:
  title: Synapse API
  version: 0.1.0
  summary: Organizational Brain service for AI agents and team workflows.
  description: |
    Synapse is the governed layer that gives agents and teams permission-aware
    retrieval, a versioned skill registry, bounded memory, a user/team/org
    context service, a policy-guarded tool/connector (MCP) gateway, and durable
    workflow orchestration.

    Production identity is carried by a verified bearer JWT (JWKS, static RS256,
    or HS256). Trusted X-Principal-Id/X-Tenant-Id headers remain available for
    development or a trusted identity-aware gateway. Every governed endpoint is
    attributable for audit.

  license:
    name: Apache-2.0
    url: https://www.apache.org/licenses/LICENSE-2.0
  contact:
    name: ResQ Software
    url: https://github.com/resq-software/synapse
servers:
  - url: http://localhost:8080
    description: Local development
...
paths:
...
  /retrieve:
    post:
      tags: [retrieval]
      operationId: retrieve
      summary: Permission-aware hybrid retrieval.
      description: |
        Runs permission-aware retrieval (hybrid vector + lexical, optional
        rerank and graph expansion) scoped by tenant, team, and ACLs. Returns a
        `trace_id` so callers can correlate logs.
      parameters:
        - $ref: '#/components/parameters/PrincipalId'
        - $ref: '#/components/parameters/TenantId'
        - $ref: '#/components/parameters/TeamIds'
        - $ref: '#/components/parameters/Role'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RetrieveRequest'
      responses:
        '200':
          description: Retrieval results.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RetrieveResponse'
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Verified JWT; required when Synapse is configured for JWT authentication.
  parameters:
    PrincipalId:
      name: X-Principal-Id
      in: header
      required: false
      description: Trusted-header identity for development/gateway mode; ignored when JWT auth is configured.
      schema:
        type: string
      example: alice@resq
    TenantId:
      name: X-Tenant-Id
      in: header
      required: false
      description: Owning tenant/organization for this request.
      schema:
        type: string
      example: resq
    TeamIds:
      name: X-Team-Ids
      in: header
      required: false
      description: Comma-separated team ids the principal is acting within.
      schema:
        type: string
      example: team-ops,team-air
    Role:
      name: X-Role
      in: header
      required: false
      description: Platform/business role of the principal, used for RBAC.
      schema:
        type: string
      example: operator
  responses:
    BadRequest:
      description: The request was malformed or failed validation.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Unauthorized:
      description: Missing or invalid bearer token or trusted principal.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Forbidden:
      description: Authenticated but not permitted by policy.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
  schemas:
    Error:
      type: object
      description: Stable error envelope returned for all failure responses.
      required: [error]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code:
              type: string
              description: Stable, machine-readable error code.
              examples: [not_found, bad_request, unauthorized, forbidden, db_error, internal]
            message:
              type: string
              description: Human-readable error message.
...
    RetrieveScope:
      type: object
      description: Team/project fences layered on top of ACLs.
      properties:
        team_ids:
          type: array
          items:
            type: string
        project_ids:
          type: array
          items:
            type: string
    RetrievalOptions:
      type: object
      description: Retrieval tuning options.
      properties:
        mode:
          type: string
          description: Retrieval mode.
          enum: [hybrid, vector, lexical]
          default: hybrid
        top_k:
          type: integer
          description: Maximum number of results to return.
          minimum: 1
          default: 10
        rerank:
          type: boolean
          description: Whether to apply a second-stage reranker.
          default: false
        include_graph:
          type: boolean
          description: Whether to include knowledge-graph expansion.
          default: false
    RetrieveRequest:
      type: object
      description: A permission-aware retrieval request.
      required: [tenant_id, principal_id, query]
      properties:
        tenant_id:
          type: string
          example: resq
        principal_id:
          type: string
          example: alice@resq
        query:
          type: string
          description: The natural-language (and/or keyword) query.
        scope:
          $ref: '#/components/schemas/RetrieveScope'
        retrieval:
          $ref: '#/components/schemas/RetrievalOptions'
    RetrievalResult:
      type: object
      description: A single retrieval hit.
      required: [chunk_id, doc_id, score, text]
      properties:
        chunk_id:
          type: string
        doc_id:
          type: string
        score:
          type: number
          format: float
          description: Relevance score for the hit.
        text:
          type: string
        source_uri:
          type: string
          description: URI of the source document, when available.
        metadata:
          type: object
          additionalProperties: true
    RetrieveResponse:
      type: object
      description: Retrieval results with a correlation trace id.
      required: [results, trace_id]
      properties:
        results:
          type: array
          items:
            $ref: '#/components/schemas/RetrievalResult'
        trace_id:
          type: string
          description: Correlation id for logs/observability.
...